[CHORE] VoteOrNot 함수 코드이동
