+++ src/api/vote.jsx
@@ -17,3 +17,14 @@ export const voting = ({ questionId, userId, votedUserIds, hint }) => {
     }
   );
 };
+
+export const voteOrNot = async (questionId) => {
+  return axios
+    .get(`${process.env.REACT_APP_BASE_URL}/votes/check/${questionId}`,{
+      headers: {
+        Authorization: "Bearer " + getCookie("refreshToken"),
+      },
+    })
+    .then(res => res.data.result);
+};
+
